fix: indent on type CR issues
