+++ server/src/sas/LanguageServiceProvider.ts
@@ -7,6 +7,7 @@ import { Model } from "./Model";
 import { CompletionProvider } from "./CompletionProvider";
 import { DocumentSymbol, SymbolKind } from "vscode-languageserver-types";
 import type { LibService } from "./SyntaxDataProvider";
+import { FormatOnTypeProvider } from "./FormatOnTypeProvider";
 
 export const legend = {
   tokenTypes: [
@@ -44,11 +45,16 @@ export class LanguageServiceProvider {
   private model;
   private syntaxProvider;
   public completionProvider;
+  public formatOnTypeProvider;
 
   constructor(doc: TextDocument) {
     this.model = new Model(doc);
     this.syntaxProvider = new SyntaxProvider(this.model);
     this.completionProvider = new CompletionProvider(
+      this.model,
+      this.syntaxProvider,
+    );
+    this.formatOnTypeProvider = new FormatOnTypeProvider(
       this.model,
       this.syntaxProvider,
     );
